requirements.txt
Django==2.2.1
django-cors-headers==3.0.1
django-datatables-view==1.18.0
django-environ==0.4.5
django-filter>=2
django-safedelete==0.5.1
django-simple-history==2.7.2
djangorestframework-simplejwt==4.4.0
djangorestframework==3.9.0
gunicorn==19.9.0
pillow==6.1.0
psycopg2-binary==2.8.4
psycopg2>=2.7,<2.8 --no-binary psycopg2
pyjwt==1.7.1
python-slugify==3.0.2